requirements.txt
pandas
numpy
scikit-learn
lightgbm
pymysql
sqlalchemy
cryptography
python-dotenv
